minor changes in db tables

+++ src/main/java/com/urlShortner/Application/Users/User.java
@@ -7,7 +7,7 @@ import javax.persistence.*;
 public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
+    private Integer id;
     @Column(name = "name")
     private String name;
     @Column(name = "email")
@@ -21,7 +21,7 @@ public class User {
 
     }
 
-    public User(Long id, String name, String email, String password, long createdAt) {
+    public User(Integer id, String name, String email, String password, long createdAt) {
         this.id = id;
         this.name = name;
         this.email = email;
@@ -29,11 +29,11 @@ public class User {
         this.createdAt = createdAt;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
